wip: trustee-script library
the goal is to make a small, JS-like language (with proper scoping and
some syntax elements borrowed from rust) to manipulate logic via
trustee.core. It should have blocks for structured proofs, use the VM,
thunks, etc. that exist, but in an interactive package that works in a
repl or in jupyter.